project nuke to transition to Lua
Hopefully this will be the last project restart.
Lua now has state bags which makes the implementation of networking infinitely easier.
Dealing non-deterministic JSON stuff is also much easier due to Lua being able to deal with a couple of missing keys and not completely crash.

Expect most of the gamemode to be done within the coming two weeks :D