feat: new, much better looking loading screen
Added in a new and really sleek loading screen complete with a suite of loading progress checking so you aren't just stuck there wondering what is happening half the time.

Unfortunately, due to an under investigation FiveM issue, the loading screen doesn't fade into the game as intended. It still looks smooth enough for the time being but once the issue is fixed, the fade in will look much nicer.